+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/DeveloperUiToolPresentation.kt
@@ -1,8 +1,13 @@
 package dev.turingcomplete.intellijdevelopertoolsplugin
 
 import com.intellij.ide.BrowserUtil
-import com.intellij.ui.dsl.builder.Row
+import com.intellij.openapi.ui.popup.Balloon
+import com.intellij.openapi.ui.popup.JBPopupFactory
+import com.intellij.ui.awt.RelativePoint
+import com.intellij.ui.dsl.builder.panel
+import com.intellij.util.ui.UIUtil
 import org.jetbrains.annotations.Nls
+import javax.swing.JComponent
 
 data class DeveloperUiToolPresentation(
   @Nls(capitalization = Nls.Capitalization.Title)
@@ -22,24 +27,38 @@ data class DeveloperUiToolPresentation(
 
   interface Description {
 
-    fun Row.buildUi()
+    fun show(parentComponent: JComponent)
   }
 
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
   private class ContextHelpDescription(private val description: String) : Description {
 
-    override fun Row.buildUi() {
-      contextHelp(description)
+    override fun show(parentComponent: JComponent) {
+      val panel = panel {
+        row {
+          label("<html>$description</html>")
+        }
+      }
+      JBPopupFactory.getInstance().createBalloonBuilder(panel)
+        .setDialogMode(true)
+        .setFillColor(UIUtil.getPanelBackground())
+        .setBlockClicksThroughBalloon(true)
+        .setRequestFocus(true)
+        .createBalloon()
+        .apply {
+          setAnimationEnabled(false)
+          show(RelativePoint.getSouthOf(parentComponent), Balloon.Position.below)
+        }
     }
   }
 
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
-  private class ExternalLinkDescription(private val title: String, private val url: String) : Description {
+  private class ExternalLinkDescription(private val url: String) : Description {
 
-    override fun Row.buildUi() {
-      link(title) { BrowserUtil.open(url) }
+    override fun show(parentComponent: JComponent) {
+      BrowserUtil.open(url)
     }
   }
 
@@ -49,6 +68,6 @@ data class DeveloperUiToolPresentation(
 
     fun contextHelp(description: String): Description = ContextHelpDescription(description)
 
-    fun externalLink(title: String, url: String): Description = ExternalLinkDescription(title, url)
+    fun externalLink(url: String): Description = ExternalLinkDescription(url)
   }
 }