+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/tool/ui/generator/MultiLineTextGenerator.kt
@@ -4,8 +4,8 @@ import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.Project
 import com.intellij.ui.dsl.builder.Align
 import com.intellij.ui.dsl.builder.Panel
-import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperUiTool
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration
+import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperUiTool
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperUiToolContext
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.DeveloperToolEditor
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.common.ValueProperty
@@ -36,6 +36,10 @@ abstract class MultiLineTextGenerator(
     }.resizableRow()
   }
 
+  override fun afterBuildUi() {
+    doGenerate()
+  }
+
   protected abstract fun generate(): String
 
   open fun Panel.buildConfigurationUi() {
@@ -49,7 +53,6 @@ abstract class MultiLineTextGenerator(
   }
 
   override fun activated() {
-    doGenerate()
     configuration.addChangeListener(parentDisposable, this)
   }