+++ src/test/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/tool/ui/converter/CliCommandConverterTest.kt
@@ -1,11 +1,10 @@
-package dev.turingcomplete.intellijdevelopertoolsplugin.developertool.converter.dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.ui.converter
+package dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.ui.converter
 
 import com.intellij.openapi.Disposable
 import com.intellij.testFramework.junit5.TestApplication
 import com.intellij.testFramework.junit5.TestDisposable
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperToolConfiguration
 import dev.turingcomplete.intellijdevelopertoolsplugin.DeveloperUiToolContext
-import dev.turingcomplete.intellijdevelopertoolsplugin._internal.tool.ui.converter.CliCommandConverter
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import java.util.*